requirements.txt
aiogram
aiohttp
python-dotenv
aiofiles 
yt-dlp
